Fix type safety issues with menu items
The following commit will port files-card-body to TypeScript, revealing
several safety issues with the menu items handling code.  Fix them up
before we do that:

 - it's difficult to reason about the types of menu items when we have
   two similar but different things: items and dividers.  Instead of
   using { type: "divider" } as a placeholder for a divider, we could
   use null, but this would have a negative impact on readabilty.
   Fortunately, with explicit typing we can form what is essentially
   discriminated union.  Move the menu code to its own TypeScript file,
   continuing the shrinking of fileActions.jsx, and add a type for menu
   items.  Now TypeScript can reason about the properties that are
   present on all non-divider items (`title` and `onClick`).  This
   prevents TypeScript from complaining that `onClick` might be
   `undefined`, for example.  The explicit typing also means that if we
   make a mistake when defining a menu item, we'll see the type error
   directly at the point of the mistake.

 - className is only defined on some items, but we do an unconditional
   concatenation of `"context-menu-option " + item.className`.  That
   means that we set the class to `context-menu-option undefined`, which
   is definitely not what we wanted

 - strictly speaking we're not allowed to pass isDisabled={undefined},
   but we're doing that for most items.  Provide a default of false.

 - not really a type issue, but strange: we're using a translated string
   as a key= on our <MenuItem/> elements when we have a unique ID.  Use
   the id instead, which is what the sidebar does.
